+++ public/crtlib.h
@@ -45,7 +45,7 @@ enum
 void Q_strnupr( const char *in, char *out, size_t size_out );
 #define Q_strlwr( in, out ) Q_strnlwr( in, out, 99999 )
 void Q_strnlwr( const char *in, char *out, size_t size_out );
-int Q_strlen( const char *string );
+size_t Q_strlen( const char *string );
 int Q_colorstr( const char *string );
 char Q_toupper( const char in );
 char Q_tolower( const char in );
